Deliver CSS that should only affect pages with Bebras questions via JS
The CSS style that adapts the quiz layout width to make sure Bebras tasks
fit in the available space is moved to JS rendered in renderer.php.
This is needed to ensure it doesn't affect pages without a Bebras task,
because the content of styles.css becomes a part of a Moodle-wide style
cache.